Fix cfg() instead of jcfg() for field usage
Three of the commands in the browser window were still using the old
location for the MorphMan fields (`morph_fields` from `cfg()`). This
commit changes those commands to use the filters in `jcfg()` instead.

It also fixes the errors occuring `massTagger.per()` and
`viewMorphemes.per()` for the case where no notes in the current
selection were in the DB/had correct filters in `jcfg()`.