Use Anki's internal conf dict for storing jcfg
This uses a dict in Anki's `mw.col.conf` dict for storing MorphMan
jcfg data instead of a JSON file next to the DB. The Anki addon writing
guide states that small config data (and I would count the jcfg config
data as small) can be stored in the `conf` dict.

The config data resides in `mw.col.conf['addons']['morphman']`. If the
`addons` dict does not yet exist, it will be created during
`initJcfg()`. I have chosen an `addon` subdir because I think it makes
sense to store addon data in a namespace separate from the 'regular'
Anki config data.

As far as I can tell, this change requires a one-time full sync to function
reliably, otherwise the config dict will be overwritten for some reason.
After a full sync, it works flawlessly for me.