Fix configs not being saved to server side
Without setting the collection as "modified" after a change, the changed
settings would not be moved over.

Please note that the `jcfg()` method can be used for assignment into the
config dict as well: `jcfg("key") = "value"`. In this case, the
collection would have to be set as modified manually afterwards.