Sentence lengths: more explicit and flexible config
Rather than name an "optimal" length L but treat all lengths [L-2,L+2]
as equally optimal, just name explicitly the min and max lengths for
that interval of equally-good lengths.

And fix the logic of the `mm_tooLong` tag to match how we prioritize
notes, only applying to notes we're actually penalizing for their
length rather than notes longer than the midpoint of the optimal
interval but which may still be within it.

Also widen the optimal interval somewhat by default,
from [2,6] to [2,8], because 6 morphemes is pretty short.