Extend locations to optionally carry weights used in frequency
This allows us to introduce external data that a given morpheme
is common, based on some external corpus, without needing to
record 1000 separate locations to get a frequency of 1000.

In looking for code that does the simple `len(locs)` frequency
count and should be updated to call the new function, found a
method that looks to be dead code (`popDb`) and a computation --
the "Calculating frequency information" step familiar from the
MorphMan recalc progress-bar dialog -- whose results are
apparently never used!  So cut both of those.

Also introduce a new Corpus location type, convenient for
this form of usage.