Remove configuration for the mecab part-of-speech blacklist
Parts of speech are a pretty blunt tool for identifying morphemes to
ignore.  It is handy that we can easily filter out punctuation, but it
doesn't seem particularly convincing as a way to filter anything else.
So move the list into the code.  This makes `morph.morphemizer` able
to be imported on its own and function completely independent of being
part of a running Anki application.

While here, we discover that (a) the bit of the blacklist that was
supposed to filter out unknown morphemes doesn't look like it could
ever have worked, because it's spelled `UNKNOWN` but the format string
we pass to mecab with `--unk-format` says `Unknown`; but (b) it
appeared to work because the unknown morphemes never made it that far,
because the same format string gave them only 4 tab-separated fields
and we filter out lines that don't have MECAB_NODE_LENGTH = 5
tab-separated fields.  So regardless of the blacklist, it was already
effectively hard-coded that unknown morphemes would be dropped.
There's a simpler way to do that -- just pass an empty format string
with `--unk-format`.  Do that.

Also erase remaining, unused traces of the part-of-speech whitelist
functionality that existed in previous versions.