Refactor how we start mecab, and add more comments and checks
This came out of some time spent studying how MorphMan starts its
`mecab` subprocess, and studying MeCab itself, in order to reproduce
the same `mecab` output outside the context of a full running Anki
application.  Hopefully it will be helpful for the next person.

One stumbling block I'd hit is that MeCab can work with a variety of
dictionaries (in fact this is something the MeCab upstream is very
proud of!), and the `mecab` I'd originally gotten on my system outside
of the Japanese Support addon was using a different dictionary
(jumandic) from the one bundled in that addon (ipadic), but our
`mecab` command line embeds specific assumptions about the dictionary.
The new `bos-feature` check catches that issue.  Handily `mecab -P`
provides both that information and the charset information we'd been
getting from `mecab -D`, so this is basically free.

This refactor pushes the logic of "run one command to get config info,
then run the real command" down underneath the logic of "try one way
to run `mecab`, then try another".  That allows us to also use the
first command to validate that the `mecab` we've found will work for
us, in particular with the `bos-feature` check just mentioned.

Also switch the order of candidate `mecab` invocations to try the
plain command line `mecab` first, and only after that go poking around
the inside of the Japanese Support addon to try to borrow its bundled
`mecab`.  That allows the user to bypass the latter logic entirely by
simply installing a working, compatible `mecab` on their machine.
Which is especially handy when using the CLI.