mecab: use both `mecab -D` and `mecab -P`, for different info
Empirically, it's possible for the charsets given by these two
commands to differ; when they do, the one given by `mecab -D` is
the one `mecab` actually communicates in.  So use `mecab -D` for
that information after all, in addition to `mecab -P` to check
the dictionary format in hopes of catching a mixup there.

Each command takes under 10ms on my machine, so the cost of running
both is negligible.

I still don't quite understand the different meanings of the two
flags; for example, on my system if I pass `-d` pointing at either
the `ipadic` or `ipadic-utf8` dictionaries (as provided by the
corresponding packages in Debian), the two charset indications
vary in tandem between `EUC-JP` and `UTF-8`.  But on @kaegi's
system, a plain `mecab -D` shows charset `utf-8` while `mecab -P`
shows `EUC-JP`, so there is a difference.