Remove `mm_badLength` tag, replace with `mm_tooShort`
This makes the tags on too-short and too-long texts much more
symmetrical and easy to describe, and provides all the same
information to filter on.  It also gets rid of the "bad length" name,
which isn't particularly clear about what it means and sounds a lot
like an error message, in favor of the very straightforward
descriptions "too short" and "too long".