Add enhanced frequency.txt support.
The old frequency handling used only the base form of the morphemes, this
could lead to some inaccuracies such as getting morphemes to learn that
look the same, but aren't the same in the database.

This new format, identified by #study_plan_frequency 1.0 at the top of
the frequency list, enables complete morpheme matching.